backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
sqlalchemy>=2.0.0
mysql-connector-python>=8.0.0
pydantic>=2.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
slowapi>=0.1.9
httpx>=0.27.0
oss2>=2.18.0
